Linode schemas
