+++ packages/api-v4/src/linodes/backups.ts
@@ -1,8 +1,6 @@
+import { CreateSnapshotSchema } from '@linode/validation/lib/linodes.schema';
 import { API_ROOT } from 'src/constants';
-
 import Request, { setData, setMethod, setURL } from '../request';
-
-import { CreateSnapshotSchema } from './linodes.schema';
 import { LinodeBackup, LinodeBackupsResponse } from './types';
 
 type Backup = LinodeBackup;
