+++ packages/api-v4/src/linodes/configs.ts
@@ -1,16 +1,16 @@
+import {
+  CreateLinodeConfigSchema,
+  UpdateLinodeConfigSchema,
+} from '@linode/validation/lib/linodes.schema';
 import { API_ROOT } from 'src/constants';
 import Request, {
   setData,
   setMethod,
   setParams,
   setURL,
-  setXFilter
+  setXFilter,
 } from '../request';
 import { ResourcePage as Page } from '../types';
-import {
-  CreateLinodeConfigSchema,
-  UpdateLinodeConfigSchema
-} from './linodes.schema';
 import { Config, LinodeConfigCreationData } from './types';
 
 /**
