+++ packages/api-v4/src/linodes/index.ts
@@ -10,8 +10,6 @@ export * from './info';
 
 export * from './ips';
 
-export * from './linodes.schema';
-
 export * from './linodes';
 
 export * from './types';
