+++ packages/manager/src/features/linodes/LinodesDetail/LinodeRebuild/LinodeRebuildDialog.tsx
@@ -5,8 +5,8 @@ import Typography from 'src/components/core/Typography';
 import EnhancedSelect, { Item } from 'src/components/EnhancedSelect/Select';
 import HostMaintenanceError from '../HostMaintenanceError';
 import LinodePermissionsError from '../LinodePermissionsError';
-import RebuildFromImage from './RebuildFromImage_CMR';
-import RebuildFromStackScript from './RebuildFromStackScript_CMR';
+import RebuildFromImage from './RebuildFromImage';
+import RebuildFromStackScript from './RebuildFromStackScript';
 import Dialog from 'src/components/Dialog';
 import useExtendedLinode from 'src/hooks/useExtendedLinode';
 import Notice from 'src/components/Notice';
