+++ packages/manager/src/features/linodes/LinodesDetail/LinodeRebuild/index.ts
@@ -1,2 +1,3 @@
-import LinodeRebuild from './LinodeRebuild';
+// @todo update this once the Bare Metal rebuild flow is merged
+import LinodeRebuild from './LinodeRebuildDialog';
 export default LinodeRebuild;
